requirements.txt
psycopg2-binary>=2.8
sqlalchemy
sqlalchemy-utils
discord.py[voice]
python-dotenv
emoji
lxml
PyNaCl
aiohttp[speedups]
coloredlogs
uvloop
